Enable a few legacy tests
- Create a base class for System.IO.Ports tests
- Break out old Support.cs into separate files

In porting the tests we need to flip methods from bool to void,
using Assert where we would have added/returned false.

Using ConditionalFact for dependencies on particular hardware
setup (with properties from the base class).

Remove all Console.WriteLines - for ease switching to Debug.WriteLine.
