+++ src/System.IO.Ports/tests/Legacy/Support/TestDelegate.cs
@@ -1,0 +1,8 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace Legacy.Support
+{
+    public delegate bool TestDelegate();
+}
